# src/config/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Cadastro de Veículos",
    "version": "1.0.0",
    "description": "API REST para gerenciamento de veículos e usuários. Autenticação por JWT e refresh tokens."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "post": {
        "summary": "Register user (creates access + refresh tokens)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
        "responses": { "201": { "description": "Created" } }
      },
      "get": {
        "summary": "List users",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login (returns access token and sets refresh cookie)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" }, "password": { "type": "string" } }, "required": ["email","password"] } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}